requirements.txt
mailjet_rest
numpy
rumps
openai
openai-whisper
py2app
pygame
PyAudio
tiktoken
pvporcupine
SpeechRecognition
google-api-core
google-cloud-texttospeech
certifi
soundfile
pipdeptree
spacy
